coverity: add a cast to truncate the time value

+++ common/integer.c
@@ -235,7 +235,7 @@ intRandInit (void)
 {
     if (intrandflag == 0) {                          /* If generator not yet initialized */
 #if ! ((defined COMMON_DEBUG) || (defined COMMON_RANDOM_FIXED_SEED) || (defined SCOTCH_DETERMINISTIC))
-        intrandseed = (unsigned int)( time(NULL) );  /* Set random seed if needed */
+        intrandseed = (unsigned int)((int)( time(NULL) ));  /* Set random seed if needed */
 #endif /* ((defined COMMON_DEBUG) || (defined COMMON_RANDOM_FIXED_SEED)) */
 #ifdef COMMON_RANDOM_RAND
         srand (intrandseed);
